Firebase Login Update 2
For pages that need to display data from firebase, need to return a FutureBuilder for a page because it is asychronous.

Reset password functionality still incomplete but can send an reset password email (given valid email).
Security Question needs to be taken from Firebase knowing if the email exists in Firebase beforehand.

Remember Me functionality needs to be reworked.

Send email verification in the future?