Firebase Login Update 4
Dealt with login error: creating a regular email/password account and then logging in with google afterwards.

Changed age field in the sign-up screen to a dropdown menu. Items in the dropdown menu may be changed if age-group ranges are clarified.

This means that for the User model, the age goes from int to String.